+++ dali/operators/decoder/decoder_test.h
@@ -64,7 +64,7 @@ class DecodeTestBase : public GenericDecoderTest<ImgType> {
     DeviceWorkspace *ws) override {
     // single input - encoded images
     // single output - decoded images
-    TensorVector<CPUBackend> out(inputs[0]->num_samples());
+    TensorList<CPUBackend> out(inputs[0]->num_samples());
     std::vector<Tensor<CPUBackend>> tmp_out(inputs[0]->num_samples());
     const TensorList<CPUBackend> &encoded_data = *inputs[0];
     const int c = this->GetNumColorComp();
