+++ dali/operators/image/remap/displacement_filter_impl_cpu.h
@@ -194,7 +194,7 @@ class DisplacementFilter<CPUBackend, Displacement, per_channel_transform>
   float fill_value_;
 
   bool has_mask_;
-  const TensorVector<CPUBackend> *mask_ = nullptr;
+  const TensorList<CPUBackend> *mask_ = nullptr;
 };
 
 }  // namespace dali
