+++ dali/pipeline/data/type_traits.h
@@ -20,7 +20,7 @@
 namespace dali {
 
 template <typename Backend>
-class TensorVector;
+class TensorList;
 
 class CPUBackend;
 class MixedBackend;
@@ -33,22 +33,19 @@ struct is_backend {
                                 std::is_same<Backend, GPUBackend>::value;
 };
 
-template <template <typename> class MaybeTensorVector, typename Backend>
-struct is_tensor_vector {
-  static constexpr bool value =
-      std::is_same<MaybeTensorVector<Backend>, TensorVector<Backend>>::value;
+template <template <typename> class MaybeTensorList, typename Backend>
+struct is_tensor_list {
+  static constexpr bool value = std::is_same<MaybeTensorList<Backend>, TensorList<Backend>>::value;
 };
 
 /**
  * Verifies, that T is proper batch container for DALI
  *
- * Batch container is proper when it has a defined backend and is TensorVector or a TensorList
+ * Batch container is proper when it has a defined backend and is TensorList or a TensorList
  */
 template <template <typename Backend_> class T, typename Backend>
 struct is_batch_container {
-  static constexpr bool value =
-      is_backend<Backend>::value &&
-      is_tensor_vector<T, Backend>::value;
+  static constexpr bool value = is_backend<Backend>::value && is_tensor_list<T, Backend>::value;
 };
 
 template <typename BatchType, typename T = void>
